bot/requirements.txt
aiogram
tortoise-orm[asyncpg]
